Move the structured_data functionalities into separate _utils module
