+++ eth_account/_utils/structured_data/__init__.py
@@ -1,0 +1,3 @@
+from .hashing import (  # noqa: F401
+    hashStruct,
+)
